[fix] fixed xack builder check on str vs int
