+++ utils/tests/test_data/api_builder_common.py
@@ -1,4 +1,4 @@
-from redis_benchmarks_specification.__api__.api import commit_schema_to_stream
+from redis_benchmarks_specification.__api__.schema import commit_schema_to_stream
 from redis_benchmarks_specification.__builder__.builder import (
     builder_consumer_group_create,
     builder_process_stream,
